pyproject.toml
[tool.poetry]
name = "posterr"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.scripts]
start = "posterr.app:main"

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8.1"
pymongo = "^4.1.1"
vyper-config = "^1.1.1"
typeguard = "^2.13.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
